The first algorithm implemented from ADM.
This is an algorithm to select the max number of jobs from a set (list) of
jobs.  The start and end times of each job is given as a list of pairs.

This is an n^2 algorithm.  Can we do better?

[In the book the author talks about selecting movies (presumably to make the
discussion more interesting/amusing), but we have generalized the term and talk
about "jobs".  It's just a change of terminology; it doesn't change the
essential nature of the problem.]